two camera

+++ Engine/TinyEngine/Other/GeometryGenerator.h
@@ -7,7 +7,7 @@ class GeometryGenerator
 public:
 	static GeometryGenerator* instance();
 public:
-	MeshComponentPtr createCubeMesh(InputLayoutType layoutType, int width, int height, int lenth);
+	MeshComponentPtr createCubeMesh(InputLayoutType layoutType, float width, float height, float lenth);
 	MeshComponentPtr createSphereMeshData(InputLayoutType layoutType, float radius, int numSubdivisions);
 protected:
 	void subdivideTriangle(std::vector<Vector3>& pos, std::vector<int>& indices);
